requirements.txt
torch
torchvision
torchaudio
scikit-learn
matplotlib
tqdm
Pillow
huggingface_hub
datasets
